fix: seperate the functions in modules
